+++ module/move/assistant/src/lib.rs
@@ -8,61 +8,9 @@
 /// Internal namespace.
 pub( crate ) mod private
 {
-
-  pub use openai_api_rs::v1::
-  {
-    api::Client,
-    assistant::AssistantObject,
-  };
-
-  use std::
-  {
-    env,
-    error::Error,
-  };
-
-  use former::Former;
-
-  /// Options for configuring the OpenAI API client.
-  #[ derive( Former, Debug ) ]
-  pub struct ClientOptions
-  {
-    /// The API key for authenticating with the OpenAI API.
-    pub api_key : Option< String >,
-  }
-
-  /// Creates a new OpenAI API client using the API key from the environment variable `OPENAI_API_KEY`.
-  pub fn client() -> Result< Client, Box< dyn Error > >
-  {
-    let api_key = env::var( "OPENAI_API_KEY" )?;
-    Ok( Client::new( api_key ) )
-  }
-
-
 }
 
-// /// Reflections.
-// pub mod reflect;
-// pub use reflect::*;
-// /// Universal wrapper.
-// pub mod wrapper_ref;
-// pub use wrapper_ref::*;
-// /// Universal wrapper.
-// pub mod wrapper_option_cow_ref;
-// pub use wrapper_option_cow_ref::*;
-// /// Conversion to string.
-// pub mod to_string;
-// pub use to_string::*;
-
-/// Nice print.
-pub mod print;
-// pub use print::*;
-/// Nice print's wrapper.
-pub mod as_table;
-// pub use as_table::*;
-/// Table interface.
-pub mod table;
-// pub use table::*;
+pub mod client;
 
 #[ allow( unused_imports ) ]
 pub use protected::*;
@@ -75,9 +23,7 @@ pub mod protected
   #[ allow( unused_imports ) ]
   pub use super::
   {
-    print::orphan::*,
-    as_table::orphan::*,
-    table::orphan::*,
+    client::orphan::*,
   };
 
   #[ doc( inline ) ]
@@ -102,18 +48,7 @@ pub mod exposed
   #[ allow( unused_imports ) ]
   pub use super::
   {
-    print::exposed::*,
-    as_table::exposed::*,
-    table::exposed::*,
-  };
-
-  #[ doc( inline ) ]
-  #[ allow( unused_imports ) ]
-  pub use super::private::
-  {
-    ClientOptions,
-    client,
-    AssistantObject,
+    client::exposed::*,
   };
 
   #[ doc( inline ) ]
@@ -135,9 +70,7 @@ pub mod prelude
   #[ allow( unused_imports ) ]
   pub use super::
   {
-    print::prelude::*,
-    as_table::prelude::*,
-    table::prelude::*,
+    client::prelude::*,
   };
 
 }
