+++ module/move/assistant/tests/inc/basic_test.rs
@@ -1,6 +1,7 @@
 #[ allow( unused_imports ) ]
 use super::*;
 
-mod basic_test;
-
-mod experiment;
+#[ test ]
+fn basic()
+{
+}
